observability-python/requirements.txt
requests==2.31.0
openai==1.50.0
python-dotenv==1.0.0
pandas==2.1.4
datetime
urllib3==2.1.0
json-logging==1.3.0
asyncio
aiohttp==3.9.1
